pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Jon Bringhurst <jon@bringhurst.org>"]
description = "Communications Hill Neighborhood Association Website"
name = "adeline"
packages = [
    { include = "adeline", from = "src" }
]
version = "0.1.0"

[tool.poetry.dependencies]
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
python = "^3.10"
Flask-Migrate = "^3.1.0"
alembic = "^1.7.4"
SQLAlchemy = "^1.4.25"
Flask-GraphQL = "^2.0.1"
graphene-sqlalchemy = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
pylint = "^2.11.1"
pytest = "^5.2"
mypy = "^0.910"
pytest-cov = "^3.0.0"
types-Flask = "^1.1.4"
flask-sqlalchemy-stubs = "^0.2"
graphene-stubs = "^0.15"
sqlalchemy-stubs = "^0.4"

[tool.poetry.scripts]
run = "adeline.webapp:main"

[tool.isort]
profile = "black"

[tool.pylint.'MESSAGES CONTROL']
disable = ["too-few-public-methods"]

[tool.mypy]
# Setup plugins
plugins = "sqlmypy"

# Import discovery
mypy_path = "src"
namespace_packages = true

# Disallow dynamic typing
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = false
disallow_subclassing_any = false

# Untyped definitions and calls
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false

# None and Optional handling
no_implicit_optional = true
strict_optional = true

# Configuring warning
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true

# Suppressing errors
show_none_errors = true
ignore_errors = false

# Miscellaneous strictness flags
allow_untyped_globals = false
allow_redefinition = false
implicit_reexport = false
strict_equality = true

# Configuring error messages
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
color_output = true
error_summary = true
show_absolute_path = true

# Advanced options
show_traceback = true
raise_exceptions = true

# Miscellaneous
warn_unused_configs = true
verbosity = 0
